test(radio): migrate item
