+++ core/src/components/radio/test/item/radio.e2e.ts
@@ -1,0 +1,39 @@
+import { expect } from '@playwright/test';
+import { configs, test } from '@utils/test/playwright';
+
+configs().forEach(({ title, screenshot, config }) => {
+  test.describe(title('radio: item'), () => {
+    test('should render correctly in list', async ({ page }) => {
+      await page.setContent(
+        `
+        <ion-list>
+          <ion-radio-group>
+            <ion-item>
+              <ion-radio>Enable Notifications</ion-radio>
+            </ion-item>
+          </ion-radio-group>
+        </ion-list>
+      `,
+        config
+      );
+      const list = page.locator('ion-list');
+      expect(await list.screenshot()).toMatchSnapshot(screenshot(`toggle-list`));
+    });
+    test('should render correctly in inset list', async ({ page }) => {
+      await page.setContent(
+        `
+        <ion-list inset="true">
+          <ion-radio-group>
+            <ion-item>
+              <ion-radio>Enable Notifications</ion-radio>
+            </ion-item>
+          </ion-radio-group>
+        </ion-list>
+      `,
+        config
+      );
+      const list = page.locator('ion-list');
+      expect(await list.screenshot()).toMatchSnapshot(screenshot(`radio-inset-list`));
+    });
+  });
+});
