+++ core/src/components/radio/test/item/radio.e2e.ts-snapshots/radio-inset-list-md-rtl-Mobile-Firefox-linux.png
